[bugfix] test hangs when two vertical-collections share container
When passive event listeners are supported, and two
vertical-collection instances share the same container element, then
the first vc instance adds an event listener, but the second vc
instance starts polling for the event.

We only want to add the event listener (once) and want to avoid
polling altogether when passive event listeners are supported.

Starting the polling pushed an extra job into the ember-raf-scheduler which was never
removed. This caused ember-raf-scheduler's testWaiter to hang indefinitly.

The fixes the issue by avoiding polling for the event, because we
already have the event listener setup, which will call all registered
handlers with the scroll event.

Fixes https://github.com/html-next/vertical-collection/issues/320